+++ assets/js/menu/submenu.ts
@@ -3,26 +3,28 @@
 import { Base } from './base'
 
 export class Submenu extends Base {
-    private timeKeeper: NodeJS.Timeout | undefined
-
     run(): void {
         this.hide()
-        // this.handleFocusOut()
+        this.handleClickOutside()
     }
 
     private hide(): void {
         this._j(this._submenu_selector).hide()
     }
 
-    private handleFocusOut(): void {
-        this._j(this._submenu_selector).parent('li').attr('tabindex', -1)
+    private handleClickOutside(): void {
+        this._j(document).on('click', (event: JQuery.ClickEvent) => {
+            if (event.isDefaultPrevented()) return
 
-        this._j(this._submenu_selector).parent('li').on('focusout', () => {
-            clearTimeout(this.timeKeeper)
+            const parent = this._j(this._submenu_selector).parent('li').get(0)
+            if (parent && this._j.contains(parent, event.target)) return
 
-            this.timeKeeper = setTimeout(() => {
-                this._j(this._submenu_selector).slideUp(this._fx_duration)
-            }, 150);
+            this._j(this._submenu_selector)
+                .prev('a')
+                .children(this._submenu_button_selector)
+                .html(this.renderIcon('down'))
+
+            this._j(this._submenu_selector).slideUp(this._fx_duration)
         })
     }
 }