+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove `fa-sm` sizing from Fontawesome icons
 - Rewrite JS for menus
 - Limit menu JS enhancement to primary menu
+- Close open submenus when a click is triggered outside it
 
 ### Removed
 - Remove `@grottopress/scss` NPM package